requirements.txt
torch>=1.8.0
torchvision>=0.9.0
transformers>=4.0.0
datasets>=2.0.0
numpy>=1.19.0
pandas>=1.2.0
matplotlib>=3.3.0
seaborn>=0.11.0
tensorboard>=2.4.0
Pillow>=8.0.0
tqdm>=4.60.0
scikit-learn>=0.24.0

